+++ CHANGES.rst
@@ -5,6 +5,7 @@ Version 0.9.1 (UNRELEASED)
 --------------------------
 
 - Changes the interactive session notification by adding a message stating that the session will be closed after a specified number of days inactivity.
+- Adds an extra button for login with custom third-party Keycloak SSO authentication services.
 
 Version 0.9.0 (2023-01-19)
 --------------------------